add all to stages to one file
