--- target_chains/cosmwasm/wormhole-stub/testnet/guardian-upgrade/osmosis_4-2.14.9.json
@@ -1,20 +1,0 @@
-{
-  "GUARDIAN_SET_UPGRADE_1_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "068EF47F42D6A1D44EFD4100B91CCE5A5050F505A088567271F5B6689A224EE5"
-    }
-  },
-  "GUARDIAN_SET_UPGRADE_2_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "AA7D06318A6945F1519A8C7DAD75141BCBC21C1F7E800B48C8045AAF1BF8A887"
-    }
-  },
-  "GUARDIAN_SET_UPGRADE_3_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "0C67CABFC942171919425A1D8A64AB40D4FF5FC49667D0B6B161704DB9C22E0C"
-    }
-  }
-}
