--- target_chains/cosmwasm/wormhole-stub/testnet/guardian-upgrade/osmosis_5-2.14.9.json
@@ -1,20 +1,0 @@
-{
-  "GUARDIAN_SET_UPGRADE_1_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "0C75CE16C91F32A902E43A6326B63800DA5182EFC52AA245E101C6374E3671B1"
-    }
-  },
-  "GUARDIAN_SET_UPGRADE_2_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "F58EF5AC1A1941362339A2355F2A2DD44BF46522C37E3D60602C0E731B36F0B6"
-    }
-  },
-  "GUARDIAN_SET_UPGRADE_3_VAA": {
-    "status": "fulfilled",
-    "result": {
-      "txHash": "59984BB3216E6A7D44501B11EE1F51735E9DE9C8D24D87343B9DDB480F3B5ED3"
-    }
-  }
-}
